refactor: remove-dbcallqueries
