+++ frontend/src/container/MetricsApplication/MetricsPageQueries/ExternalQueries.tsx
@@ -228,6 +228,107 @@ export const externalCallDurationByAddress = ({
 	],
 });
 
+// DBCall Queries
+
+export const databaseCallsRPS = ({
+	servicename,
+	legend,
+	tagFilterItems,
+}: DatabaseCallsRPSProps): {
+	formulas: IMetricsBuilderFormula[];
+	queryBuilder: IMetricsBuilderQuery[];
+} => ({
+	formulas: [],
+	queryBuilder: [
+		{
+			aggregateOperator: 18,
+			disabled: false,
+			groupBy: ['db_system'],
+			legend,
+			metricName: 'signoz_db_latency_count',
+			name: 'A',
+			reduceTo: 1,
+			tagFilters: {
+				items: [
+					{
+						id: '',
+						key: 'service_name',
+						op: 'IN',
+						value: [`${servicename}`],
+					},
+					...tagFilterItems,
+				],
+				op: 'AND',
+			},
+		},
+	],
+});
+
+export const databaseCallsAvgDuration = ({
+	servicename,
+	tagFilterItems,
+}: ExternalCallProps): {
+	formulas: IMetricsBuilderFormula[];
+	queryBuilder: IMetricsBuilderQuery[];
+} => ({
+	formulas: [
+		{
+			disabled: false,
+			expression: 'A/B',
+			name: 'F1',
+			legend: '',
+		},
+	],
+	queryBuilder: [
+		{
+			aggregateOperator: 18,
+			disabled: true,
+			groupBy: [],
+			legend: '',
+			metricName: 'signoz_db_latency_sum',
+			name: 'A',
+			reduceTo: 1,
+			tagFilters: {
+				items: [
+					{
+						id: '',
+						key: 'service_name',
+						op: 'IN',
+						value: [`${servicename}`],
+					},
+					...tagFilterItems,
+				],
+				op: 'AND',
+			},
+		},
+		{
+			aggregateOperator: 18,
+			disabled: true,
+			groupBy: [],
+			legend: '',
+			metricName: 'signoz_db_latency_count',
+			name: 'B',
+			reduceTo: 1,
+			tagFilters: {
+				items: [
+					{
+						id: '',
+						key: 'service_name',
+						op: 'IN',
+						value: [`${servicename}`],
+					},
+					...tagFilterItems,
+				],
+				op: 'AND',
+			},
+		},
+	],
+});
+
+interface DatabaseCallsRPSProps extends ExternalCallProps {
+	legend: '{{db_system}}';
+}
+
 interface ExternalCallDurationByAddressProps extends ExternalCallProps {
 	legend: '{{address}}';
 }
